+++ src/commands/change-issue-status.ts
@@ -29,7 +29,7 @@ export class ChangeIssueStatusCommand implements Command {
         }
       } else {
         if (canExecuteJiraAPI()) {
-          throw new Error('Use this command from JIRA: EXPLORER');
+          printErrorMessageInOutput('Use this command from Jira Plugin EXPLORER');
         }
       }
     } catch (err) {